Structurize project according to common style
Structuring of a project makes it a lot easier to navigate through
project files. Arrange current project files according to the style set
in Ruby projects - https://github.com/ruby
