+++ ruby/doc/git.txt
@@ -1,0 +1,15 @@
+GIT
+
+mkdir /root/python/practice
+cd /root/python/practice
+git clone https://github.com/belgeu/misc.git
+cd misc
+## edit .git/config to ask pass (add belgeu@ to url)
+## url = https://belgeu@github.com/belgeu/misc.git
+git add webdriver_test.rb
+git commit -m "MESSAGE"
+git push origin master
+
+DOCs:
+- Selenium Finding elements
+https://gist.github.com/huangzhichong/3284966
